+++ src/index.js
@@ -2,6 +2,7 @@ import * as animators from './animators/index';
 import * as react from './react/index';
 import animate from './animate';
 import getRandomConfig from './get-random-config';
+import drawCanvas from './draw-canvas';
 
-export { animators, react, animate, getRandomConfig };
-export default { animators, react, animate, getRandomConfig };
+export { animators, react, animate, getRandomConfig, drawCanvas };
+export default { animators, react, animate, getRandomConfig, drawCanvas };